AUI-1394 Create tutorial for Affix
Note: the jira ticket is to create both a tutorial and examples for this
module. This commit only contains the tutorial part. The api link is not
yet on the site's api, but should be correct. The "Configuring Affix"
section needs the most work -- not clear yet if there are other options,
and if this section should be removed if there are not.