AUI-1400 Create tutorial for Dropdown
Note: the jira ticket is to create both a tutorial and examples for this
module. This commit only contains the tutorial part. The linked api does
not exist at this point, but it should be the correct location. There is a
way to make a dropdown menu without using the 'items' attribute (using a
nested <ul> element instead) that is included in the demo, but it relies
on particular classes, so is not included in this tutorial.