AUI-1398 Create tutorial for Undo/Redo
Note: the jira ticket is to create both a tutorial and examples for this
module. This commit only contains the tutorial part. The "Using Undo/Redo"
section needs the most attention -- I used the demo for reference and am
not sure if it actually optional to define new undo and redo methods or if
it is necessary. Right now it is presented as optional.