Fixing a few examples
(1) Carousel - We're using the `image-viewer-base-image` class to construct from markup now, instead of `carousel-item`.
(2) Pagination - The CSS classes should be in different places now.
(3) Modal - The problem here seems to be a bug on the current implementation of alloy's modal, when setting both toolbar and height. I removed height from this demo so it would be better at least until we fix this.